scaffolds/python-fullstack/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
